+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 1.0.1
 
 - Change log and add verbose option.
+- Display runtime after download ends.
 
 ## 1.0.0